modules/flutter_i18n/auto_translate/requirements.txt
googletranslate_python==1.0.3
PyYAML==6.0.1
